pyproject.toml
[tool.poetry]
name = "dds"
version = "0.1.0"
description = ""
authors = [" annavaleriev <filenko.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.1"
djangorestframework = "^3.16.0"
django-filter = "^25.1"
drf-spectacular = "^0.28.0"
psycopg2-binary = "^2.9.10"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 119

